integration
- Moved smaller modules into `dank.tool.py`, they should now launch even faster!
